requirements.txt
django==4.2.13
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
drf-yasg==1.21.7
Pygments==2.18.0
psycopg2-binary==2.9.8
djangorestframework-simplejwt==5.3.1
